+++ test/MemoryPool/clear.cpp
@@ -2,15 +2,16 @@
 // Copyright Benoit Blanchon 2014-2018
 // MIT License
 
-#include <ArduinoJson/Memory/StaticMemoryPool.hpp>
+#include <ArduinoJson/Memory/MemoryPool.hpp>
 #include <catch.hpp>
 
 using namespace ARDUINOJSON_NAMESPACE;
 
 static const size_t poolCapacity = 512;
 
-TEST_CASE("StaticMemoryPool::clear()") {
-  StaticMemoryPool<poolCapacity> memoryPool;
+TEST_CASE("MemoryPool::clear()") {
+  char buffer[poolCapacity];
+  MemoryPool memoryPool(buffer, sizeof(buffer));
 
   SECTION("Discards allocated variants") {
     memoryPool.allocVariant();
@@ -21,6 +22,7 @@ TEST_CASE("StaticMemoryPool::clear()") {
 
   SECTION("Discards allocated strings") {
     memoryPool.allocFrozenString(10);
+    REQUIRE(memoryPool.size() > 0);
 
     memoryPool.clear();