+++ test/MemoryPool/size.cpp
@@ -2,44 +2,61 @@
 // Copyright Benoit Blanchon 2014-2018
 // MIT License
 
-#include <ArduinoJson/Memory/StaticMemoryPool.hpp>
+#include <ArduinoJson/Memory/MemoryPool.hpp>
 #include <catch.hpp>
 
 using namespace ARDUINOJSON_NAMESPACE;
 
-TEST_CASE("StaticMemoryPool::size()") {
-  SECTION("Capacity equals template parameter") {
-    const size_t capacity = 64;
-    StaticMemoryPool<capacity> memoryPool;
-    REQUIRE(capacity == memoryPool.capacity());
-  }
+char buffer[4096];
+
+TEST_CASE("MemoryPool::capacity()") {
+  const size_t capacity = 64;
+  MemoryPool memoryPool(buffer, capacity);
+  REQUIRE(capacity == memoryPool.capacity());
+}
+
+TEST_CASE("MemoryPool::size()") {
+  MemoryPool memoryPool(buffer, sizeof(buffer));
 
   SECTION("Initial size is 0") {
-    StaticMemoryPool<32> memoryPool;
     REQUIRE(0 == memoryPool.size());
   }
 
+  SECTION("size() == capacity() after allocExpandableString()") {
+    memoryPool.allocExpandableString();
+    REQUIRE(memoryPool.size() == memoryPool.capacity());
+  }
+
+  SECTION("Decreases after freezeString()") {
+    StringSlot* a = memoryPool.allocExpandableString();
+    memoryPool.freezeString(a, 1);
+    REQUIRE(memoryPool.size() == JSON_STRING_SIZE(1));
+
+    StringSlot* b = memoryPool.allocExpandableString();
+    memoryPool.freezeString(b, 1);
+    REQUIRE(memoryPool.size() == 2 * JSON_STRING_SIZE(1));
+  }
+
   SECTION("Increases after allocFrozenString()") {
-    StaticMemoryPool<128> memoryPool;
     memoryPool.allocFrozenString(0);
     REQUIRE(memoryPool.size() == JSON_STRING_SIZE(0));
+
     memoryPool.allocFrozenString(0);
     REQUIRE(memoryPool.size() == 2 * JSON_STRING_SIZE(0));
   }
 
   SECTION("Decreases after freeVariant()") {
-    StaticMemoryPool<128> memoryPool;
     VariantSlot* a = memoryPool.allocVariant();
     VariantSlot* b = memoryPool.allocVariant();
 
     memoryPool.freeVariant(b);
     REQUIRE(memoryPool.size() == sizeof(VariantSlot));
+
     memoryPool.freeVariant(a);
     REQUIRE(memoryPool.size() == 0);
   }
 
   SECTION("Decreases after calling freeString() in order") {
-    StaticMemoryPool<128> memoryPool;
     StringSlot* a = memoryPool.allocFrozenString(5);
     REQUIRE(a != 0);
     StringSlot* b = memoryPool.allocFrozenString(6);
@@ -52,7 +69,6 @@ TEST_CASE("StaticMemoryPool::size()") {
   }
 
   SECTION("Decreases after calling freeString() in reverse order") {
-    StaticMemoryPool<128> memoryPool;
     StringSlot* a = memoryPool.allocFrozenString(5);
     REQUIRE(a != 0);
     StringSlot* b = memoryPool.allocFrozenString(6);
@@ -65,15 +81,13 @@ TEST_CASE("StaticMemoryPool::size()") {
   }
 
   SECTION("Doesn't grow when memory pool is full") {
-    const size_t variantCount = 4;
-    const size_t capacity = variantCount * sizeof(VariantSlot);
-    StaticMemoryPool<capacity> memoryPool;
+    const size_t variantCount = sizeof(buffer) / sizeof(VariantSlot);
 
     for (size_t i = 0; i < variantCount; i++) memoryPool.allocVariant();
-    REQUIRE(capacity == memoryPool.size());
+    size_t size = memoryPool.size();
 
     memoryPool.allocVariant();
 
-    REQUIRE(capacity == memoryPool.size());
+    REQUIRE(size == memoryPool.size());
   }
 }